demo: avoid implicit conversions
Issue #23

+++ repos/demo/include/libpng_static/pngconf.h
@@ -1082,7 +1082,7 @@
  * want to have unsigned int for png_uint_32 instead of unsigned long.
  */
 
-typedef unsigned long png_uint_32;
+typedef unsigned int png_uint_32;
 typedef long png_int_32;
 typedef unsigned short png_uint_16;
 typedef short png_int_16;
